crawling/requirements.txt
playwright
requests
beautifulsoup4
lxml
supabase
python-dotenv
